+++ app/components/MainLayout.js
@@ -3,7 +3,7 @@ import React, { Component, PropTypes } from 'react'
 import mainRoutes from '../routing/mainRoutes'
 
 import SplitPane from 'react-split-pane'
-import ServerView from '../containers/serverView'
+import ServerView from '../containers/ServerView'
 import { SERVER_VIEW_MIN_SIZE, RIGHT_PANEL_MIN_SIZE } from '../constants/defaultSettings'
 
 export default class Main extends Component {
@@ -20,15 +20,15 @@ export default class Main extends Component {
   render () {
     return (
       <SplitPane
-        className="main-layout-container"
-        split="vertical"
+        className='main-layout-container'
+        split='vertical'
         minSize={SERVER_VIEW_MIN_SIZE}
         defaultSize={this.props.serverViewWidth}
         onChange={this.props.serverViewWidthChanged}
         maxSize={-RIGHT_PANEL_MIN_SIZE}
       >
         <ServerView />
-        <div className="right-panel">
+        <div className='right-panel'>
           {mainRoutes}
         </div>
       </SplitPane>
